chore: refactor sign uniform intervals iterator.
