+++ playgrounds/terrain/src/sdf/analysis/interval/sign_uniform_intervals/intervals.rs
@@ -1,7 +1,9 @@
 pub mod boundary_mapping;
+pub mod interval_interator;
 pub mod interval_mapping;
 
 use crate::sdf::analysis::interval::{SignBoundary, SignUniformInterval};
+use interval_interator::SignUniformIntervalsIterator;
 use std::collections::BTreeSet;
 
 #[derive(Debug, Clone, Default, PartialEq, Eq)]
@@ -37,62 +39,13 @@ impl SignUniformIntervals {
 	}
 }
 
-pub struct SignUniformIntervalsIterator {
-	intervals: Vec<SignBoundary>,
-	index: usize,
-	emitted_top: bool,
-}
-
-// Iterates left, right pairs beginning with the top constant, then through the members of the set, then ending with the bottom constant.
-impl Iterator for SignUniformIntervalsIterator {
-	type Item = SignUniformInterval;
-	fn next(&mut self) -> Option<Self::Item> {
-		if self.intervals.is_empty() {
-			if !self.emitted_top {
-				self.emitted_top = true;
-				return Some(SignUniformInterval {
-					left: SignBoundary::top(),
-					right: SignBoundary::bottom(),
-				});
-			}
-			return None;
-		}
-
-		if !self.emitted_top {
-			self.emitted_top = true;
-			// First pair: (top, first_element)
-			return Some(SignUniformInterval {
-				left: SignBoundary::top(),
-				right: self.intervals[0].clone(),
-			});
-		}
-
-		if self.index < self.intervals.len() - 1 {
-			// Middle pairs: (elem_i, elem_{i+1})
-			let left = self.intervals[self.index].clone();
-			let right = self.intervals[self.index + 1].clone();
-			self.index += 1;
-			return Some(SignUniformInterval { left, right });
-		}
-
-		if self.index < self.intervals.len() {
-			// Last pair: (last_element, bottom)
-			let left = self.intervals[self.index].clone();
-			self.index += 1;
-			return Some(SignUniformInterval { left, right: SignBoundary::bottom() });
-		}
-
-		None
-	}
-}
-
 impl IntoIterator for SignUniformIntervals {
 	type Item = SignUniformInterval;
 	type IntoIter = SignUniformIntervalsIterator;
 
 	fn into_iter(self) -> Self::IntoIter {
 		let intervals: Vec<SignBoundary> = self.boundaries.into_iter().collect();
-		SignUniformIntervalsIterator { intervals, index: 0, emitted_top: false }
+		SignUniformIntervalsIterator::new(intervals)
 	}
 }
